# api/openapi.yaml
openapi: '3.1.0'
info:
  title: apikit API
  version: 0.1.0
  description: >-
    The apikit API provides user management, organization management,
    authentication, and authorization services.
servers:
  - url: /api/v1
    description: Default mount point (configurable per deployment)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Bearer token authentication supporting three credential types:
        Admin Token (format: <prefix>_admin_<64 hex chars>) for break-glass
        emergency access with global scope;
        API Key (format: <prefix>_<key_id>_<secret>) for user-scoped full
        access to own resources;
        PAT (format: <prefix>_pat_<token_id>_<secret>) for fine-grained
        per-permission access.
  headers:
    XRequestID:
      description: UUID for distributed tracing, present on every response
      schema:
        type: string
        format: uuid
...
    CacheControlNoStore:
      description: Cache-Control directive for mutable resources
      schema:
        type: string
        enum:
          - 'no-store'
...
  schemas:
    User:
      type: object
      description: A user resource representing an authenticated account.
      required:
        - id
        - username
        - email
        - status
        - role
        - provider
        - provider_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user.
        username:
          type: string
          description: Display name for the user, set at creation and immutable via PATCH.
        email:
          type: string
          description: Email address for the user, set at creation and immutable via PATCH.
        full_name:
          type:
            - string
            - 'null'
          description: Optional display name for the user. The only field updatable via PATCH.
        status:
          type: string
          enum:
            - active
            - blocked
          description: Current status of the user account.
        role:
          type: string
          enum:
            - admin
            - user
          description: Access level of the user.
        provider:
          type: string
          description: OAuth identity provider name (e.g. github, google).
        provider_id:
          type: string
          description: Unique user identifier assigned by the OAuth provider.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the user was created (RFC 3339, UTC with Z suffix).
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the user was last modified (RFC 3339, UTC with Z suffix).
...
    Error:
      type: object
      description: >-
        Standard error envelope. The code field equals the HTTP response
        status code exactly.
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: integer
              description: >-
                The HTTP status code of the response. Mirrors the HTTP status
                exactly — no additional error codes are used.
            message:
              type: string
              description: A human-readable error description.
    AuthCallbackResponse:
      type: object
      description: Response from the OAuth callback endpoint containing user and API key.
      required:
        - user
        - api_key
      properties:
        user:
          $ref: '#/components/schemas/User'
        api_key:
          type: object
          required:
            - key
            - key_id
            - expires_at
          properties:
            key:
              type: string
              description: Full API key string including the plaintext secret.
            key_id:
              type: string
              description: The random 8-character alphanumeric identifier portion of the API key.
            expires_at:
              type:
                - string
                - 'null'
              format: date-time
              description: >-
                When the API key expires (RFC 3339 UTC). Null when expires=0
                (permanent, no expiry).
...
  responses:
    BadRequest:
      description: Bad request — invalid or malformed request body.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    PayloadTooLarge:
      description: Payload Too Large — the request body exceeds the configured size limit.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnsupportedMediaType:
      description: Unsupported Media Type — Content-Type must be application/json.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /auth/callback:
    post:
      operationId: postAuthCallback
      summary: OAuth callback
      description: >-
        Exchanges an OAuth authorization code for a user record and API key.
        Public — no authentication required. Request body must use
        Content-Type: application/json or receive 415.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
                - code
                - redirect_uri
              properties:
                provider:
                  type: string
                  description: OAuth provider name (e.g. github, google).
                code:
                  type: string
                  description: OAuth authorization code.
                redirect_uri:
                  type: string
                  description: OAuth redirect URI for validation.
                expires:
                  type: integer
                  enum:
                    - 0
                    - 30
                    - 60
                    - 90
                  default: 90
                  description: >-
                    Expiry in days for the issued API key. 0 means no expiry
                    (permanent); expires_at will be null. 30, 60, or 90 means
                    the key expires exactly 24h × N days from creation. Default
                    is 90 days.
      responses:
        '200':
          description: OAuth callback successful. Returns user and API key.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthCallbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: The matched user account is blocked.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The provider-supplied email is already registered to a different account.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '502':
          description: >-
            The provider user info request failed, or the provider returned an
            empty user id.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...